Restore binary compatibility with 3.7.5

+++ cpp/src/Ice/InputStream.cpp
@@ -608,12 +608,12 @@ Ice::InputStream::read(pair<const Short*, const Short*>& v, IceUtil::ScopedArray
         v.second = reinterpret_cast<Short*>(i);
 #else
 #  ifdef ICE_CPP11_MAPPING
-        auto result = new short[sz];
+        auto result = new short[static_cast<size_t>(sz)];
         _deleters.push_back([result] { delete[] result; });
         v.first = result;
         v.second = result + sz;
 #  else
-        result.reset(new Short[sz]);
+        result.reset(new Short[static_cast<size_t>(sz)]);
         v.first = result.get();
         v.second = result.get() + sz;
 #   endif
@@ -691,12 +691,12 @@ Ice::InputStream::read(pair<const Int*, const Int*>& v, ::IceUtil::ScopedArray<I
 #else
 
 #  ifdef ICE_CPP11_MAPPING
-        auto result = new int[sz];
+        auto result = new int[static_cast<size_t>(sz)];
         _deleters.push_back([result] { delete[] result; });
         v.first = result;
         v.second = result + sz;
 #  else
-        result.reset(new Int[sz]);
+        result.reset(new Int[static_cast<size_t>(sz)]);
         v.first = result.get();
         v.second = result.get() + sz;
 #  endif
@@ -812,12 +812,12 @@ Ice::InputStream::read(pair<const Long*, const Long*>& v, IceUtil::ScopedArray<L
 #else
 
 #  ifdef ICE_CPP11_MAPPING
-        auto result = new long long[sz];
+        auto result = new long long[static_cast<size_t>(sz)];
         _deleters.push_back([result] { delete[] result; });
         v.first = result;
         v.second = result + sz;
 #  else
-        result.reset(new Long[sz]);
+        result.reset(new Long[static_cast<size_t>(sz)]);
         v.first = result.get();
         v.second = result.get() + sz;
 #  endif
@@ -925,12 +925,12 @@ Ice::InputStream::read(pair<const Float*, const Float*>& v, IceUtil::ScopedArray
 #else
 
 #  ifdef ICE_CPP11_MAPPING
-        auto result = new float[sz];
+        auto result = new float[static_cast<size_t>(sz)];
         _deleters.push_back([result] { delete[] result; });
         v.first = result;
         v.second = result + sz;
 #  else
-        result.reset(new Float[sz]);
+        result.reset(new Float[static_cast<size_t>(sz)]);
         v.first = result.get();
         v.second = result.get() + sz;
 #  endif
@@ -1046,12 +1046,12 @@ Ice::InputStream::read(pair<const Double*, const Double*>& v, IceUtil::ScopedArr
 #else
 
 #  ifdef ICE_CPP11_MAPPING
-        auto result = new double[sz];
+        auto result = new double[static_cast<size_t>(sz)];
         _deleters.push_back([result] { delete[] result; });
         v.first = result;
         v.second = result + sz;
 #  else
-        result.reset(new Double[sz]);
+        result.reset(new Double[static_cast<size_t>(sz)]);
         v.first = result.get();
         v.second = result.get() + sz;
 #  endif
